Change Main component into an object
In this situation the Main acts as a controller but it isn't a Component in sense that it isn't organized in a main.reel folder with its own template.

Usually a component template is considered an internal detail of the component itself, this is why they are organized into .reel directories. It is the role of the component to define what its template is supposed to be.

In this case the main object has no "influence" over the template file (index.html) it controls, it is the index.html that includes the main object instead of the other way around.

A future suggestion is to transform the main object into a proper component by placing it into ui/main.reel along with index.html renamed to main.html.

Since main is no longer a component it doesn't draw and prepareForDraw isn't called. We can still use the templateDidLoad delegate method (that is invoked on all "owner" objects of a template) to prepare the UI.