Use InputText and event details to add users
One of the things I try to avoid is accessing the DOM inside the JavaScript object, this is because when we do it we're creating a DOM dependency from the object to the template.

Here I define an InputText component in the template so if in the JavaScript I need to access its value I use the component and not the DOM element. In the future it will be easier to give the HTML file to a designer, things will still work even if she/he changes the entire DOM structure.

But there is no need to access the InputText because we're able to define the information the Button sends with its "action" event.
We create a binding from the InputText.value to the Button.detail.username.
When the button dispatches an action event its detail object will be available in the event object as well.