Use condition components to show and hide elements
There's a component in Montage that allows the developer to show and hide elements. This way we can separate further the business logic from the UI logic.

We create a state in the main object that indicates if it is loading and then we bind that state to the Condition.condition property. When condition is true its contents are visible and when it is false its contents are invisible.

The two conditions created are "loadingCondition" that controls the "Loading..." text and "contentCondition" that controls the entire content.

"loadingCondition" just needs to bind directly to the owner.loading property but for "contentCondition" we want the reverse logic, so we use an InvertConverter that does exactly this. With FRB (the future bindings of Montage) this should be as simple as defining a binding with "!@owner.loading".

Now the main object is only responsible to maintain the state of its loading property and doesn't need to worry with how that will influence the UI, that logic resides in the template.

If in the future we need to add more UI artifacts related to the loading state we can just bind to that property and the main object does not need to be changed for something that is UI related.

The "montage-invisible" is a class that the Condition sets on the element when it is hidden, in order to avoid a flash I set it manually on the content because it is its initial state. This is only a problem with elements that are in the index.html, the browser will render the HTML before the JavaScript is able to kick in.